requirements.txt
django > 3.2
reportlab
django-crispy-forms
docutils
djangorestframework
